remove generate key code duplication
